fix(qa): cycle 7 bugs discovered by tests — hasBranches + FK ordering
- Message.hasBranches: Int64 cast (was Int, always false due to GRDB type)
- TreeStore.deleteBranch: capture sessionId first, delete branch before
  session (FK-safe order prevents constraint violation)

Both bugs found by the new test suite (90 tests, 0 failures).

+++ Sources/Core/Models/Message.swift
@@ -73,7 +73,7 @@ extension Message: FetchableRecord {
             createdAt = Date()
         }
 
-        hasBranches = (row["has_branches"] as? Int ?? 0) > 0
+        hasBranches = (row["has_branches"] as? Int64 ?? 0) > 0
     }
 }
